+++ src/pages/Detail/index.tsx
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from 'react';
 import {
   Flex,
   Container,
-  Box,
   Text,
   Image,
   Button,
@@ -11,7 +10,6 @@ import {
   AccordionButton,
   AccordionPanel,
   AccordionIcon,
-  Link,
   Spinner,
   Modal,
   ModalOverlay,
@@ -22,45 +20,19 @@ import { RouteComponentProps, useHistory, Link as RouterLink } from 'react-route
 import axios from 'axios';
 import qs from 'qs';
 import MainContainer from '../../layout/MainContainer';
-import PriceHistoryChart from './PriceHistoryChart';
 import CancelDialog from './CancelDialog';
 import DealDialog from './DealDialog';
-import NoData from './NoData';
+import DetailLeft from './DetailLeft';
+import DetailRight from './DetailRight';
 import { getBlock } from '../../polkaSDK/api/getBlock';
-import TimeBy from './TimeBy';
-import { renderNmtNumberText } from '../../components/Balance';
 import {
   CACHE_SERVER_URL,
   PINATA_SERVER,
 } from '../../constants';
 
 import {
-  IconDetailsocllections,
-  // IconDetailsRefresh,
-  // IconDetailshaSre,
-  IconDetailsCollection,
-  IconDetailsCollectionS,
-  IconDetailsDetail,
-  WEBSITE,
-  DISCORD,
-  TWITTER,
-  IconIns,
-  medium,
-  telegram,
-  // PriceIcon,
   Historyempty,
-  HeadPortrait,
-  IconAbout,
-  IconPriceHistory,
-  IconProperties,
-  Iconprice,
   IconOffersDetail,
-  ImgFillTop,
-  ImgFillBottom,
-  IconAuthentication,
-  IconDetailsRefresh,
-  IconDetailshaSre,
-  // IconLeft,
 } from '../../assets/images';
 import useNft from '../../hooks/reactQuery/useNft';
 import useCollectionsSinger from '../../hooks/reactQuery/useCollectionsSinger';
@@ -73,15 +45,6 @@ import BuyDialog from './BuyDialog';
 import OfferDialog from './OfferDialog';
 
 // const propertiesArr = [1, 2, 3, 4, 5, 6];
-const ICONS = [
-  { icon: WEBSITE.default, name: 'website' },
-  { icon: DISCORD.default, name: 'discord' },
-  { icon: TWITTER.default, name: 'twitter' },
-  { icon: IconIns.default, name: 'ins' },
-  { icon: medium.default, name: 'medium' },
-  { icon: telegram.default, name: 'telegram' },
-];
-
 const OfferssUnitArr = [1, 2, 3, 4, 5, 6];
 
 const Detail = ({ match }: RouteComponentProps<{ nftId: string }>) => {
@@ -96,7 +59,6 @@ const Detail = ({ match }: RouteComponentProps<{ nftId: string }>) => {
   const { data: nftData, isLoading: nftDataIsLoading } = useNft(nftId);
   const collectionsId = nftId.split('-')[0];
   const tokenId = nftId.split('-')[1];
-  const formatAddress = (addr: string) => `${addr.slice(0, 4)}...${addr.slice(-4)}`;
 
   const collectNft = async (type:string) => {
     const data = {
@@ -132,7 +94,6 @@ const Detail = ({ match }: RouteComponentProps<{ nftId: string }>) => {
   const [isShowBuy, setIsShowBuy] = useState(false);
   const [isShowOffer, setIsShowOffer] = useState(false);
   const [isSubmitting, setIsSubmitting] = useState(false);
-  const [selectedTime, setSelectedTime] = useState('seven');
   const [offerId, setOfferId] = useState('');
   const [offerOwner, setOfferOwner] = useState('');
   useEffect(() => {
@@ -144,77 +105,20 @@ const Detail = ({ match }: RouteComponentProps<{ nftId: string }>) => {
   }, []);
 
   const { data: token } = useToken();
-  const { data: collectionsData, isLoading: collectionsDateIsLoading } = useCollectionsSinger(collectionsId);
   const isLoginAddress = useIsLoginAddress(nftData?.nftInfo.owner_id);
+  const { data: collectionsData, isLoading: collectionsDateIsLoading } = useCollectionsSinger(collectionsId);
   if (nftDataIsLoading || collectionsDateIsLoading || !nftData) {
     return <Spinner />;
   }
-  const links = collectionsData?.collection?.metadata?.links;
-  const ICON_LIST = ICONS.map((item, index) => ({
-    src: item.icon,
-    id: index,
-    link: links ? links[item.name] : '',
-  }));
 
   const logoUrl = `${PINATA_SERVER}${nftData.nftInfo.metadata.logoUrl}`;
   const price = priceStringDivUnit(nftData?.nftInfo?.price);
   const collectionName = collectionsData?.collection?.metadata?.name;
-  const collectionDescription = collectionsData?.collection?.metadata?.description;
   const nftName = nftData?.nftInfo?.metadata.name;
-  const OffersArr = nftData?.nftInfo?.offers;
 
   const ownerId = nftData?.nftInfo?.owner_id;
   const orderId = nftData?.nftInfo?.order_id;
   const hideFlag = false;
-  const PriceHistory = nftData?.nftInfo?.history[selectedTime];
-
-  const handleDeal = (offerIdItem:string, offerOwnerItem:string) => {
-    if (!account) {
-      history.push(`/connect?callbackUrl=item/${nftId}`);
-    }
-    setOfferId(offerIdItem);
-    setOfferOwner(offerOwnerItem);
-    setIsShowDeal(true);
-  };
-  const handleBuy = () => {
-    if (!account) {
-      history.push(`/connect?callbackUrl=item/${nftId}`);
-    }
-    setIsShowBuy(true);
-  };
-  const handleOffer = () => {
-    if (!account) {
-      history.push(`/connect?callbackUrl=item/${nftId}`);
-    }
-    setIsShowOffer(true);
-  };
-
-  const timeBlock = (index:numer) => {
-    const times = (index - remainingTime) * 6;
-
-    let theTime = parseInt(times.toString(), 10);
-    let middle = 0;
-    let hour = 0;
-
-    if (theTime > 60) {
-      middle = parseInt((theTime / 60).toString(), 10);
-      theTime = parseInt((theTime % 60).toString(), 10);
-      if (middle > 60) {
-        hour = parseInt((middle / 60).toString(), 10);
-        middle = parseInt((middle % 60).toString(), 10);
-      }
-    }
-    // let result = null;
-    let result = `${parseInt(theTime.toString(), 10)}`;
-    if (middle > 0) {
-      result = `${parseInt(middle.toString(), 10)}:${result}`;
-    }
-    if (hour > 0) {
-      result = `${parseInt(hour.toString(), 10)}:${result}`;
-      // result = `${parseInt(hour.toString(), 10)}`;
-    }
-    return result;
-  };
 
   return (
     <MainContainer title={`${nftName}-${collectionName}${t('Detail.title')}`}>
@@ -280,7 +184,6 @@ const Detail = ({ match }: RouteComponentProps<{ nftId: string }>) => {
                 </Flex>
 
               </Flex>
-
             )
             : (
               <Flex
@@ -339,999 +242,21 @@ const Detail = ({ match }: RouteComponentProps<{ nftId: string }>) => {
           justifyContent="space-between"
           alignItems="flex-start"
         >
-          <Flex width="560px" flexDirection="column">
-            <Flex
-              width="560px"
-              display="flex"
-              justifyContent="center"
-              alignItems="center"
-            >
-              <Box
-                display="flex"
-                flexDirection="column"
-                alignItems="center"
-                position="relative"
-              >
-                <Image
-                  width="40px"
-                  height="40px"
-                  src={ImgFillTop.default}
-                  position="absolute"
-                  right="0px"
-                />
-                <Image
-                  m="20px"
-                  maxWidth="520px"
-                  height="auto"
-                  src={logoUrl}
-                />
-                <Image
-                  position="absolute"
-                  left="0px"
-                  bottom="0px"
-                  width="40px"
-                  height="40px"
-                  src={ImgFillBottom.default}
-                />
-              </Box>
-            </Flex>
-            <Flex className="DetailLeft" width="560px">
-              <Accordion width="560px" defaultIndex={[0, 1, 2]} allowMultiple>
-                <AccordionItem width="100%" border="none">
-                  <AccordionButton
-                    height="62px"
-                    width="100%"
-                    display="flex"
-                    justifyContent="space-between"
-                    alignItems="center"
-                    p="0 20px"
-                    borderBottom="1px solid #E5E5E5"
-                    outline="none"
-                    _focus={{
-                      textDecoration: 'none',
-                      boxShadow: 'none',
-                    }}
-                  >
-                    <Flex height="100%" alignItems="center">
-                      <Image
-                        mr="8px"
-                        w="22px"
-                        h="22px"
-                        src={IconDetailsDetail.default}
-                      />
-                      <Text
-                        fontSize="16px"
-                        fontFamily="TTHoves-Medium, TTHoves"
-                        fontWeight="500"
-                        color="#000000"
-                        lineHeight="18px"
-                      >
-                        {t('Detail.detail')}
-                      </Text>
-                    </Flex>
-                    <AccordionIcon />
-                  </AccordionButton>
-                  <AccordionPanel p="16px 20px">
-                    <Flex alignItems="center" mb="11px">
-                      <Image
-                        mr="8px"
-                        w="22px"
-                        h="22px"
-                        src={HeadPortrait.default}
-                      />
-                      <Text
-                        fontSize="14px"
-                        fontFamily="TTHoves-Regular, TTHoves"
-                        fontWeight="400"
-                        color="#000000"
-                        lineHeight="16px"
-                      >
-                        <Text
-                          fontSize="14px"
-                          fontFamily="TTHoves-Regular, TTHoves"
-                          fontWeight="400"
-                          color="#000000"
-                          display="flex"
-                          flexDirection="row"
-                        >
-                          {t('Detail.createdBy')}
-                          <Link
-                            as={RouterLink}
-                            to={`/account/${nftData?.nftInfo?.creator_id}/wallet`}
-                            m="0 3px"
-                            fontSize="14px"
-                            fontFamily="TTHoves-Regular, TTHoves"
-                            fontWeight="400"
-                            color="#3D00FF"
-                          >
-                            {nftData ? formatAddress(nftData?.nftInfo?.creator_id) : ''}
-                          </Link>
-                          {/* 2018-09-2 */}
-                        </Text>
-                      </Text>
-                    </Flex>
-                    <Text
-                      fontSize="14px"
-                      fontFamily="TTHoves-Light, TTHoves"
-                      fontWeight="300"
-                      color="#000000"
-                      lineHeight="22px"
-                    >
-                      {nftData?.nftInfo?.metadata.description}
-                    </Text>
-                  </AccordionPanel>
-                </AccordionItem>
-                {/* <AccordionItem width="100%" border="none">
-                <AccordionButton
-                  height="62px"
-                  width="100%"
-                  display="flex"
-                  justifyContent="space-between"
-                  alignItems="center"
-                  p="0 20px"
-                  borderBottom="1px solid #E5E5E5"
-                  outline="none"
-                  _focus={{
-                    textDecoration: 'none',
-                    boxShadow: 'none',
-                  }}
-                >
-                  <Flex height="100%" alignItems="center">
-                    <Image
-                      mr="8px"
-                      w="22px"
-                      h="22px"
-                      src={IconProperties.default}
-                    />
-                    <Text
-                      fontSize="16px"
-                      fontFamily="TTHoves-Medium, TTHoves"
-                      fontWeight="500"
-                      color="#000000"
-                      lineHeight="18px"
-                    >
-                      {t('Detail.properties')}
-                    </Text>
-                  </Flex>
-                  <AccordionIcon />
-                </AccordionButton>
-                {hideFlag
-                  ? (
-                    <AccordionPanel
-                      p="16px 20px 4px 20px"
-                      display="flex"
-                      flexFlow="row wrap"
-                      justifyContent="space-between"
-                    >
-                      {propertiesArr.map((item) => (
-                        <Flex
-                          key={item}
-                          mb="12px"
-                          width="165px"
-                          height="86px"
-                          borderRadius="4px"
-                          border="1px solid #000000"
-                          flexDirection="column"
-                          alignItems="center"
-                          justifyContent="center"
-                        >
-
-                          <Text
-                            fontSize="12px"
-                            fontFamily="TTHoves-Regular, TTHoves"
-                            fontWeight="400"
-                            color="#000000"
-                            lineHeight="14px"
-                          >
-                            HEAD
-                          </Text>
-                          <Text
-                            m="8px 0"
-                            fontSize="14px"
-                            fontFamily="TTHoves-Medium, TTHoves"
-                            fontWeight="400"
-                            color="#000000"
-                            lineHeight="18px"
-                          >
-                            PANDA
-                          </Text>
-                          <Text
-                            fontSize="12px"
-                            fontFamily="TTHoves-Regular, TTHoves"
-                            fontWeight="400"
-                            color="#999999"
-                            lineHeight="14px"
-                          >
-                            1.23% have this trait
-                          </Text>
-                        </Flex>
-                      ))}
-
-                    </AccordionPanel>
-                  )
-                  : (
-                    <AccordionPanel p="0px">
-                      <Flex
-                        width="100%"
-                        height="260px"
-                        background="#FFFFFF"
-                        flexDirection="column"
-                        justifyContent="center"
-                        alignItems="center"
-                      >
-                        <Image
-                          w="150px"
-                          h="100px"
-                          border="1px solid #999999"
-                          borderStyle="dashed"
-                          src={Historyempty.default}
-                        />
-                        <Text
-                          mt="10px"
-                          fontSize="14px"
-                          fontFamily="TTHoves-Regular, TTHoves"
-                          fontWeight="400"
-                          color="#999999"
-                          lineHeight="20px"
-                        >
-                          No trading data yet
-                        </Text>
-                      </Flex>
-                    </AccordionPanel>
-                  )}
-
-              </AccordionItem> */}
-                <AccordionItem width="100%" border="none">
-                  <AccordionButton
-                    height="62px"
-                    width="100%"
-                    display="flex"
-                    justifyContent="space-between"
-                    alignItems="center"
-                    p="0 20px"
-                    borderBottom="1px solid #E5E5E5"
-                    outline="none"
-                    _focus={{
-                      textDecoration: 'none',
-                      boxShadow: 'none',
-                    }}
-                  >
-                    <Flex height="100%" alignItems="center">
-                      <Image
-                        mr="8px"
-                        w="22px"
-                        h="22px"
-                        src={IconAbout.default}
-                      />
-                      <Text
-                        fontSize="16px"
-                        fontFamily="TTHoves-Medium, TTHoves"
-                        fontWeight="500"
-                        color="#000000"
-                        lineHeight="18px"
-                      >
-                        {t('Detail.about')}
-                        {' '}
-                        {collectionName}
-                      </Text>
-                    </Flex>
-                    <AccordionIcon />
-                  </AccordionButton>
-                  {collectionDescription
-                    ? (
-                      <AccordionPanel p="16px 20px 16px 20px">
-                        <Text
-                          mb="16px"
-                          fontSize="14px"
-                          fontFamily="TTHoves-Light, TTHoves"
-                          fontWeight="300"
-                          color="#000000"
-                          lineHeight="22px"
-                        >
-                          {collectionDescription}
-                        </Text>
-
-                        <Flex>
-                          {ICON_LIST.map((item) => (
-                            item.link === '' ? null
-                              : (
-                                <Link
-                                  href={item.link}
-                                >
-                                  <Box
-                                    key="index"
-                                    width="40px"
-                                    height="40px"
-                                    borderRadius="4px 0px 0px 4px"
-                                    border="1px solid #E5E5E5"
-                                    display="flex"
-                                    justifyContent="center"
-                                    alignItems="center"
-                                    _hover={{
-                                      boxShadow: '0px 2px 8px 0px #E1E1E1',
-                                    }}
-                                  >
-                                    <Image
-                                      w="22px"
-                                      h="22px"
-                                      src={item.src}
-                                    />
-                                  </Box>
-
-                                </Link>
-                              )
-                          ))}
-                        </Flex>
-
-                      </AccordionPanel>
-                    )
-                    : (
-                      <NoData widths="100%" />
-                    )}
-                </AccordionItem>
-              </Accordion>
-            </Flex>
-          </Flex>
-          <Flex width="788px" flexDirection="column">
-            <Flex
-              width="788px"
-              flexDirection="column"
-              justifyContent="flex-start"
-              borderBottom="1px solid #000000"
-            >
-
-              <Flex p="0 20px 0 20px" width="100%" h="40px" justifyContent="space-between" alignItems="flex-start">
-                <Link
-                  as={RouterLink}
-                  to={`/collection/${account?.address}?collectionId=${collectionsData?.collection?.id}`}
-                >
-                  <Flex h="18px" alignItems="flex-start">
-                    <Text
-                      fontSize="14px"
-                      fontFamily="TTHoves-Regular, TTHoves"
-                      fontWeight="400"
-                      color="#3D00FF"
-                    >
-                      {collectionName}
-                    </Text>
-                    <Image
-                      ml="4px"
-                      w="18px"
-                      h="18px"
-                      src={IconAuthentication.default}
-                    />
-                  </Flex>
-                </Link>
-                <Flex>
-                  <Box
-                    width="40px"
-                    height="40px"
-                    borderRadius="4px 0px 0px 4px"
-                    border="1px solid #E5E5E5"
-                    display="flex"
-                    justifyContent="center"
-                    alignItems="center"
-                    _hover={{
-                      boxShadow: '0px 2px 8px 0px #E1E1E1',
-                    }}
-                  >
-                    <Image
-                      w="22px"
-                      h="22px"
-                      src={IconDetailsRefresh.default}
-                    />
-                  </Box>
-                  <Box
-                    width="40px"
-                    height="40px"
-                    borderTop="1px solid #E5E5E5"
-                    borderBottom="1px solid #E5E5E5"
-                    display="flex"
-                    justifyContent="center"
-                    alignItems="center"
-                    _hover={{
-                      boxShadow: '0px 2px 8px 0px #E1E1E1',
-                    }}
-                  >
-                    <Image
-                      w="22px"
-                      h="22px"
-                      src={IconDetailshaSre.default}
-                    />
-                  </Box>
-                  <Box
-                    width="40px"
-                    height="40px"
-                    borderRadius="0px 4px 4px 0px"
-                    border="1px solid #E5E5E5"
-                    display="flex"
-                    justifyContent="center"
-                    alignItems="center"
-                    _hover={{
-                      boxShadow: '0px 2px 8px 0px #E1E1E1',
-                    }}
-                    onClick={() => {
-                      setIsCollect(!isCollect);
-                      const type = isCollect ? 'cancle' : 'collect';
-                      collectNft(type);
-                    }}
-                  >
-                    <Image
-                      w="22px"
-                      h="22px"
-                      src={isCollect ? IconDetailsCollectionS.default : IconDetailsCollection.default}
-                    />
-                  </Box>
-
-                </Flex>
-              </Flex>
-              <Flex p="0 20px 0 20px">
-                <Text
-                  fontSize="40px"
-                  fontFamily="TTHoves-Bold, TTHoves"
-                  fontWeight="bold"
-                  color="#333333"
-                >
-                  {nftName}
-                </Text>
-              </Flex>
-              <Flex p="0 20px 0 20px" m="16px 0px 32px 0px" h="22px" justifyContent="flex-start" alignItems="center">
-                <Flex alignItems="center" mr="30px">
-                  <Image
-                    mr="8px"
-                    w="22px"
-                    h="22px"
-                    src={HeadPortrait.default}
-                  />
-                  <Text
-                    display="flex"
-                    flexDirection="row"
-                    fontSize="14px"
-                    fontFamily="TTHoves-Regular, TTHoves"
-                    fontWeight="400"
-                    color="#000000"
-                    lineHeight="16px"
-                  >
-                    {t('Detail.ownedBy')}
-                    <Link
-                      ml="3px"
-                      as={RouterLink}
-                      to={`/account/${nftData?.nftInfo?.owner_id}/wallet`}
-                      color="#3D00FF"
-                    >
-                      {nftData ? formatAddress(nftData?.nftInfo?.owner_id) : ''}
-                    </Link>
-                  </Text>
-                </Flex>
-                {/* <Text
-                fontSize="14px"
-                fontFamily="TTHoves-Regular, TTHoves"
-                fontWeight="400"
-                color="#000000"
-                mr="20px"
-              >
-                Listed Art
-              </Text>
-              <Flex h="22px" justifyContent="flex-start" alignItems="center" mr="20px">
-                <Image
-                  mr="4px"
-                  w="16px"
-                  h="16px"
-                  src={IconDetailsCollection.default}
-                />
-                <Text
-                  fontSize="14px"
-                  fontFamily="TTHoves-Regular, TTHoves"
-                  fontWeight="400"
-                  color="#000000"
-                >
-                  12,993
-                </Text>
-              </Flex>
-              <Flex h="22px" justifyContent="flex-start" alignItems="center">
-                <Image
-                  mr="4px"
-                  w="16px"
-                  h="16px"
-                  src={IconDetailsCollection.default}
-                />
-                <Text
-                  fontSize="14px"
-                  fontFamily="TTHoves-Regular, TTHoves"
-                  fontWeight="400"
-                  color="#000000"
-                >
-                  56
-                </Text>
-              </Flex> */}
-              </Flex>
-              {/* <Flex
-              w="100%"
-              height="40px"
-              background="#F14E1D"
-              borderRadius="4px"
-              alignItems="center"
-              pl="20px"
-            >
-              <Flex alignItems="center" mr="30px">
-                <Image
-                  mr="8px"
-                  w="16px"
-                  h="16px"
-                  src={IconDetailsCollection.default}
-                />
-                <Text
-                  fontSize="14px"
-                  fontFamily="TTHoves-Regular, TTHoves"
-                  fontWeight="400"
-                  color="#FFFFFF"
-                >
-                  Sale ends in 4 days (2021-4-1 14:32:12)
-                </Text>
-              </Flex>
-            </Flex> */}
-              <Flex
-                w="100%"
-                height="136px"
-                justifyContent="space-between"
-                alignItems="flex-start"
-                p="0 20px 0 20px"
-              >
-                <Flex h="100%" flexDirection="column" justifyContent="center">
-                  <Text
-                    fontSize="14px"
-                    fontFamily="TTHoves-Regular, TTHoves"
-                    fontWeight="400"
-                    color="#999999"
-                    lineHeight="16px"
-                    mb="12px"
-                  >
-                    {t('Detail.currentPrice')}
-                  </Text>
-                  <Flex flexDirection="column" justifyContent="flex-start">
-                    <Flex height="43px" flexDirection="row" alignItems="flex-end">
-                      <Text
-                        fontSize="36px"
-                        fontFamily="TTHoves-Bold, TTHoves"
-                        fontWeight="bold"
-                        color="#333333"
-                        lineHeight="43px"
-                      >
-                        {Number(price) ? price : '-'}
-                      </Text>
-                      <Text
-                        m="0 0 8px 4px"
-                        fontSize="14px"
-                        fontFamily="TTHoves-Regular, TTHoves"
-                        fontWeight="400"
-                        color="#999999"
-                      >
-                        {Number(price) ? `NMT ($${token?.price * Number(price)})` : null}
-                      </Text>
-                      {/* <Image
-                      m="0 0 8px 10px"
-                      w="20px"
-                      h="20px"
-                      src={IconDetailsCollection.default}
-                    /> */}
-                    </Flex>
-                  </Flex>
-                </Flex>
-                <Flex h="100%" flexDirection="column" justifyContent="center">
-                  <Button
-                    width="184px"
-                    height="48px"
-                    background={!isLoginAddress && Number(price) > 0 ? '#000000' : '#999999'}
-                    borderRadius="4px"
-                    fontSize="16px"
-                    fontFamily="TTHoves-Regular, TTHoves"
-                    fontWeight="500"
-                    color={!isLoginAddress && Number(price) > 0 ? '#FFFFFF' : '#FFFFFF'}
-                    isDisabled={isLoginAddress || Number(price) === 0}
-                    onClick={handleBuy}
-                  >
-                    {!isLoginAddress && Number(price) > 0 ? t('Detail.buyNow') : '-'}
-                  </Button>
-
-                  {/* <Text
-                  fontSize="14px"
-                  fontFamily="TTHoves-Regular, TTHoves"
-                  fontWeight="400"
-                  color="#999999"
-                >
-                  * Extend 30 minutes after bidding
-                </Text> */}
-                </Flex>
-              </Flex>
-              {/* <Flex
-              width="788px"
-              height="65px"
-              background="#F9F9F9"
-              flexDirection="row"
-              justifyContent="space-between"
-              alignItems="center"
-              p="0 20px 0 20px"
-              m="20px 0 32px 0px"
-            >
-              <Text
-                fontSize="14px"
-                fontFamily="TTHoves-Regular, TTHoves"
-                fontWeight="400"
-                color="#999999"
-              >
-                {!isLoginAddress ? '*You can buy this item immediately using Fixed Price' : '-'}
-              </Text>
-
-              <Flex flexDirection="row" justifyContent="flex-end" alignItems="center">
-                <Text
-                  fontSize="14px"
-                  fontFamily="TTHoves-Regular, TTHoves"
-                  fontWeight="bold"
-                  color="#999999"
-                  display="flex"
-                  flexDirection="row"
-                >
-                  {t('Detail.Fixedprice')}
-                  <Text
-                    m="0 3px"
-                    fontSize="14px"
-                    fontFamily="TTHoves-Regular, TTHoves"
-                    fontWeight="400"
-                    color="#000000"
-                  >
-                    {price}
-                  </Text>
-                  NMT
-                </Text>
-                {!isLoginAddress
-                  && (
-                    <Button
-                      ml="40px"
-                      width="109px"
-                      height="35px"
-                      background="#FFFFFF"
-                      borderRadius="4px"
-                      border="1px solid #3D00FF"
-                      fontSize="14px"
-                      fontFamily="TTHoves-Regular, TTHoves"
-                      fontWeight="400"
-                      color="#3D00FF"
-                      onClick={handleBuy}
-                    >
-                      {t('Detail.BuyNow')}
-                    </Button>
-                  )}
-              </Flex>
-
-            </Flex> */}
-            </Flex>
-
-            <Flex className="DetailRight" width="788px">
-              <Accordion width="100%" defaultIndex={[0, 1]} allowMultiple>
-                <AccordionItem width="100%" border="none">
-                  <AccordionButton
-                    height="62px"
-                    width="100%"
-                    display="flex"
-                    justifyContent="space-between"
-                    alignItems="center"
-                    p="0 20px"
-                    borderBottom="1px solid #E5E5E5"
-                    outline="none"
-                    _focus={{
-                      textDecoration: 'none',
-                      boxShadow: 'none',
-                    }}
-                  >
-                    <Flex height="100%" alignItems="center">
-                      <Image
-                        mr="8px"
-                        w="22px"
-                        h="22px"
-                        src={IconPriceHistory.default}
-                      />
-                      <Text
-                        fontSize="16px"
-                        fontFamily="TTHoves-Medium, TTHoves"
-                        fontWeight="500"
-                        color="#000000"
-                        lineHeight="18px"
-                      >
-                        {t('Detail.priceHistory')}
-                      </Text>
-                    </Flex>
-                    <AccordionIcon />
-                  </AccordionButton>
-
-                  <AccordionPanel p="20px">
-
-                    <Flex flexDirection="row" justifyContent="flex-start" mb="20px">
-                      <TimeBy selectedTime={selectedTime} setSelectedTime={setSelectedTime} />
-                      <Flex m="0 20px" textAlign="center" flexDirection="column" justifyContent="center">
-                        <Text
-                          mb="2px"
-                          fontSize="12px"
-                          fontFamily="TTHoves-Regular, TTHoves"
-                          fontWeight="400"
-                          color="#999999"
-                          lineHeight="14px"
-                        >
-                          {t(`Detail.${selectedTime}`)}
-                          {' '}
-                          {t('Detail.average')}
-                        </Text>
-                        <Flex align="flex-start" alignItems="center">
-                          <Box w="14px" h="14px" src={IconDetailsDetail.default} as="img" alt="" mr="4px" />
-                          <Text
-                            fontSize="16px"
-                            fontFamily="TTHoves-Regular, TTHoves"
-                            fontWeight="400"
-                            color="#000000"
-                            lineHeight="18px"
-                          >
-                            {renderNmtNumberText(PriceHistory?.average) || '-'}
-                          </Text>
-                        </Flex>
-                      </Flex>
-                      <Flex textAlign="center" flexDirection="column" justifyContent="center">
-                        <Text
-                          mb="2px"
-                          fontSize="12px"
-                          fontFamily="TTHoves-Regular, TTHoves"
-                          fontWeight="400"
-                          color="#999999"
-                          lineHeight="14px"
-                        >
-                          {t(`Detail.${selectedTime}`)}
-                          {' '}
-                          {t('Detail.volume')}
-                        </Text>
-                        <Flex align="flex-start" alignItems="center">
-                          <Box w="14px" h="14px" src={IconDetailsDetail.default} as="img" alt="" mr="4px" />
-                          <Text
-                            fontSize="16px"
-                            fontFamily="TTHoves-Regular, TTHoves"
-                            fontWeight="400"
-                            color="#000000"
-                            lineHeight="18px"
-                          >
-                            {PriceHistory?.volume || '-'}
-                          </Text>
-                        </Flex>
-                      </Flex>
-                    </Flex>
-                    {PriceHistory?.price_list.length
-                      ? (
-                        <PriceHistoryChart PriceDate={PriceHistory.price_list} />)
-                      : (
-                        <NoData widths="100%" />
-                      )}
-
-                  </AccordionPanel>
-
-                </AccordionItem>
-                <AccordionItem width="100%" border="none">
-                  <AccordionButton
-                    height="62px"
-                    width="100%"
-                    display="flex"
-                    justifyContent="space-between"
-                    alignItems="center"
-                    p="0 20px"
-                    borderBottom="1px solid #E5E5E5"
-                    outline="none"
-                    _focus={{
-                      textDecoration: 'none',
-                      boxShadow: 'none',
-                    }}
-                  >
-                    <Flex height="100%" alignItems="center">
-                      <Image
-                        mr="8px"
-                        w="22px"
-                        h="22px"
-                        src={Iconprice.default}
-                      />
-                      <Text
-                        fontSize="16px"
-                        fontFamily="TTHoves-Medium, TTHoves"
-                        fontWeight="500"
-                        color="#000000"
-                        lineHeight="18px"
-                      >
-                        {t('Detail.offers')}
-                      </Text>
-                    </Flex>
-                    <AccordionIcon />
-                  </AccordionButton>
-                  <AccordionPanel p="0 20px">
-                    <Flex w="100%" flexDirection="column" justifyContent="flex-start">
-                      {OffersArr.length
-                        ? (
-                          <>
-                            <Flex h="40px" w="100%" flexDirection="row" justifyContent="space-between" align="center">
-                              <Text
-                                w="136px"
-                                textAlign="left"
-                                fontSize="12px"
-                                fontFamily="TTHoves-Regular, TTHoves"
-                                fontWeight="400"
-                                color="#999999"
-                                lineHeight="20px"
-                              >
-                                {t('Detail.from')}
-                              </Text>
-                              <Text
-                                w="136px"
-                                textAlign="center"
-                                fontSize="12px"
-                                fontFamily="TTHoves-Regular, TTHoves"
-                                fontWeight="400"
-                                color="#999999"
-                                lineHeight="20px"
-                              >
-                                {t('Detail.price')}
-                              </Text>
-                              <Text
-                                w="136px"
-                                textAlign="center"
-                                fontSize="12px"
-                                fontFamily="TTHoves-Regular, TTHoves"
-                                fontWeight="400"
-                                color="#999999"
-                                lineHeight="20px"
-                              >
-                                {t('Detail.expiration')}
-                              </Text>
-                              <Text
-                                w="136px"
-                                textAlign="right"
-                                fontSize="12px"
-                                fontFamily="TTHoves-Regular, TTHoves"
-                                fontWeight="400"
-                                color="#999999"
-                                lineHeight="20px"
-                              />
-
-                            </Flex>
-                            <Box height="162px" overflowY="scroll" boxSizing="border-box">
-                              {OffersArr.map((item) => (
-                                <Box
-                                  key={item.id}
-                                  h="54px"
-                                >
-                                  <Flex
-                                    w="100%"
-                                    flexDirection="row"
-                                    justifyContent="space-between"
-                                    align="center"
-                                  >
-                                    <Text
-                                      w="136px"
-                                      textAlign="left"
-                                      fontSize="14px"
-                                      fontFamily="TTHoves-Regular, TTHoves"
-                                      fontWeight="400"
-                                      color="#000000"
-                                      lineHeight="20px"
-                                    >
-                                      {item?.order?.seller_id
-                                        ? formatAddress(item?.order?.seller_id)
-                                        : formatAddress(item?.order?.buyer_id)}
-                                    </Text>
-                                    <Text
-                                      w="136px"
-                                      display="flex"
-                                      flexDirection="row"
-                                      justifyContent="center"
-                                      fontSize="14px"
-                                      fontFamily="TTHoves-Regular, TTHoves"
-                                      fontWeight="400"
-                                      color="#000000"
-                                      lineHeight="20px"
-                                    >
-                                      {priceStringDivUnit(item?.order?.price)}
-                                      <Text
-                                        ml="3px"
-                                        color="#999999"
-                                      >
-                                        NMT
-                                      </Text>
-                                    </Text>
-                                    {item?.order?.deadline && item.order.status_id === 'Created' ? (
-                                      <Text
-                                        minW="136px"
-                                        textAlign="center"
-                                        fontSize="14px"
-                                        fontFamily="TTHoves-Regular, TTHoves"
-                                        fontWeight="400"
-                                        color="#000000"
-                                        lineHeight="20px"
-                                      >
-                                        in
-                                        {' '}
-                                        {timeBlock(item?.order?.deadline)}
-                                        {' '}
-                                        hours
-                                      </Text>
-                                    ) : (
-                                      <Text
-                                        w="136px"
-                                        textAlign="center"
-                                        fontSize="14px"
-                                        fontFamily="TTHoves-Regular, TTHoves"
-                                        fontWeight="400"
-                                        color="#000000"
-                                        lineHeight="20px"
-                                      >
-                                        -
-                                      </Text>
-                                    )}
-                                    {item?.order?.deadline && isLoginAddress && item.order.status_id === 'Created' ? (
-                                      <Text
-                                        w="136px"
-                                        textAlign="right"
-                                        fontSize="14px"
-                                        fontFamily="TTHoves-Regular, TTHoves"
-                                        fontWeight="400"
-                                        color="#3D00FF"
-                                        lineHeight="20px"
-                                        onClick={() => {
-                                          handleDeal(item.order_id, item.order.buyer_id);
-                                        }}
-                                      >
-                                        {' '}
-                                        {t('Detail.deal')}
-                                      </Text>
-                                    ) : (
-                                      <Text
-                                        w="136px"
-                                        textAlign="right"
-                                        fontSize="14px"
-                                        fontFamily="TTHoves-Regular, TTHoves"
-                                        fontWeight="400"
-                                        lineHeight="20px"
-                                      >
-                                        {' '}
-                                        -
-                                      </Text>
-                                    )}
-                                  </Flex>
-                                </Box>
-                              ))}
-                            </Box>
-                          </>
-                        )
-                        : (
-                          <NoData widths="732px" />
-                        )}
-                      <Flex justifyContent="flex-end">
-                        {isLoginAddress ? (
-                          ''
-                        ) : (
-                          <Button
-                            mt="16px"
-                            width="132px"
-                            height="40px"
-                            background="#FFFFFF"
-                            borderRadius="4px"
-                            border="1px solid #000000"
-                            fontSize="16px"
-                            fontFamily="TTHoves-Regular, TTHoves"
-                            fontWeight="500"
-                            color="#000000"
-                            onClick={handleOffer}
-                          >
-                            {t('Detail.makeOffer')}
-                          </Button>
-                        )}
-                      </Flex>
-                    </Flex>
-                  </AccordionPanel>
-                </AccordionItem>
-              </Accordion>
-            </Flex>
-          </Flex>
+          <DetailLeft nftData={nftData} collectionsData={collectionsData} logoUrl={logoUrl} />
+          <DetailRight
+            nftData={nftData}
+            collectionsData={collectionsData}
+            account={account}
+            isLoginAddress={isLoginAddress}
+            isCollect={isCollect}
+            remainingTime={remainingTime}
+            setOfferId={setOfferId}
+            setOfferOwner={setOfferOwner}
+            setIsShowDeal={setIsShowDeal}
+            setIsShowBuy={setIsShowBuy}
+            token={token}
+            setIsShowOffer={setIsShowOffer}
+          />
         </Flex>
         <Flex width="100%" justifyContent="space-around" />
         <Flex justifyContent="flex-end">
